don't know what i did :)

+++ backend/src/main/java/com/crimeinvestigation/system/model/Feedback.java
@@ -15,7 +15,7 @@ public class Feedback {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "feedback_id")
-    private int feedbackID;
+    private Long feedbackID;
 
     /*@ManyToOne
     @JoinColumn(crimeName = "userid")
