+++ backend/src/main/java/com/crimeinvestigation/system/model/TrackingStatus.java
@@ -5,7 +5,6 @@ import jakarta.persistence.*;
 import lombok.*;
 
 
-
 @Entity
 @Getter
 @Setter
@@ -16,7 +15,7 @@ public class TrackingStatus {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "tracking_id")
-    private int trackingID;
+    private Long trackingID;
     @ManyToOne
     @JoinColumn(name = "caseID", referencedColumnName = "caseID")
     private CrimeCase caseID;
